Through Section 6.2

+++ src/color.hpp
@@ -4,12 +4,12 @@
 #include <glm/vec3.hpp>
 #include <iostream>
 
-using color = glm::vec3;
+using color = glm::dvec3;
 
 void write_color(std::ostream &out, color pixel_color) {
-	out << int(255.999 * pixel_color.r) << ' '
-	    << int(255.999 * pixel_color.g) << ' '
-	    << int(255.999 * pixel_color.b) << '\n';
+	out << static_cast<int>(255.999 * pixel_color.r) << ' '
+	    << static_cast<int>(255.999 * pixel_color.g) << ' '
+	    << static_cast<int>(255.999 * pixel_color.b) << '\n';
 }
 
 #endif
